Crosswalked with npn
Now our data is combined in a data frame with the npn data before the loop determining gdd5 is run. This made it a lot tighter and so that one loop accomplishes the majrority of the work. Currently our trees are just given the arbs general lat and long but this could be refined.