Added visualization and better location integration
Added some visualizations of the prediction vs observed although it's not a good one. I need to figure out how to best show this when involving multiple points. Will probably talk to CHristy but since that beyond the functioning of this model, I'll leave it as is for the Bayesian transition