Improved linear model and data download
Improved the linear model by adding another level of hierarchy. Want to discuss with CHristy before I add more. I also made the main download script have the functionality of the linear one and chilling days for the alternating model so we only need one download script for the arb and one for NPN (NPN script not yet updated)