Changed covariate calculation loop to have its own output of weather data
So after cleaning up this loop for work with Andrew, I realized how I could clean it up and then thought I could have the loop produce a full data frame of weather data (similar to what we do with arb data) since pulling the daymet data is often a time constraint. For how long it took and how weird a  code it became I dont know if it was worth it.