Broke into 4 panels so each can have unique information when clicked
This is the first pass deconstructing the old plot that was one cowplot of four plots. This made it very difficult to allow for clicking to give the information desired so I've now made them each their own plot. They don't look good right now but since I'm working under the hood right now I'm more concerned with it working. Aesthetics are next