I had to make these changes in order for the script to work but I'm not sure I choose the right object
It's clear that in recent changes a naming convention was changed and there is no longer a general "ghcn.ens" but multiple such as "ghcn.ens.cfs", "ghcn.ens.gefs", and "ghcn.ens.all". Potentially the values I replaced were meant to be replaced with "ghcn.ens.all". Need Christy to check